add other oses

+++ src/os/restart.sh
@@ -7,14 +7,18 @@ cd "$(dirname "${BASH_SOURCE[0]}")" \
 
 main() {
 
-    print_in_purple "\n • Restart\n\n"
+	if [ ! -d "/mnt/c/Windows" ]; then
 
-    ask_for_confirmation "Do you want to restart?"
-    printf "\n"
+	    print_in_purple "\n • Restart\n\n"
 
-    if answer_is_yes; then
-        sudo shutdown -r now &> /dev/null
-    fi
+	    ask_for_confirmation "Do you want to restart?"
+	    printf "\n"
+
+	    if answer_is_yes; then
+	        sudo shutdown -r now &> /dev/null
+	    fi
+
+	fi
 
  }
