+++ src/os/set_github_ssh_key.sh
@@ -75,12 +75,23 @@ set_github_ssh_key() {
 
     # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
-    generate_ssh_keys "$sshKeyFileName"
-    add_ssh_configs "$sshKeyFileName"
-    copy_public_ssh_key_to_clipboard "${sshKeyFileName}.pub"
-    open_github_ssh_page
-    test_ssh_connection \
-        && rm "${sshKeyFileName}.pub"
+    local os=""
+    os="$(get_os)"
+
+    if [ "$os" == "windows" ]; then
+        print_warning "Not generating keys for Ubuntu on Windows."
+    elif [ "$os" == "ubuntu-server" ]; then
+        print_warning "Not generating keys for server OS."
+    else
+
+        generate_ssh_keys "$sshKeyFileName"
+        add_ssh_configs "$sshKeyFileName"
+        copy_public_ssh_key_to_clipboard "${sshKeyFileName}.pub"
+        open_github_ssh_page
+        test_ssh_connection \
+            && rm "${sshKeyFileName}.pub"
+
+    fi
 
 }
